# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Renoly v2 API
  version: 0.1.0
  description: |
    Machine-readable contract for the clean-slate Renoly v2 API.
    The normative prose remains docs/api-spec.md. Employee operations always
    verify the path organization, active membership, role and resource tenant.
    Cross-tenant resources are returned as 404 to avoid existence disclosure.
  contact:
    name: Renoly engineering
servers:
  - url: /api/v2
    description: Same-origin application API
...
paths:
...
  /organizations/{orgId}/assignments/{assignmentId}:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/AssignmentId"
    patch:
      tags: [Work orders]
      operationId: updateAssignment
      summary: Update assignment duty
      security: *employeeWriteSecurity
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssignmentRequest"
      responses:
        <<: *standardErrors
        "200":
          description: Assignment updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentEnvelope"
    delete:
      tags: [Work orders]
      operationId: cancelAssignment
      summary: Cancel an assignment while retaining its history
      security: *employeeWriteSecurity
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelAssignmentRequest"
      responses:
        <<: *standardErrors
        "200":
          description: Assignment cancelled; the cancelled assignment is returned.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentEnvelope"
  /organizations/{orgId}/assignments/{assignmentId}/actions/respond:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/AssignmentId"
      - $ref: "#/components/parameters/IdempotencyKey"
    post:
      tags: [Work orders]
      operationId: respondToAssignment
      summary: Accept or decline the current technician's own assignment
      security: *employeeWriteSecurity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RespondAssignmentRequest"
      responses:
        <<: *standardErrors
        "200":
          description: Assignment response recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentEnvelope"
...
  /organizations/{orgId}/work-order-checklists/{id}/actions/complete:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/ResourceId"
      - $ref: "#/components/parameters/IfMatch"
      - $ref: "#/components/parameters/IdempotencyKey"
    post:
      tags: [Checklists]
      operationId: completeWorkOrderChecklist
      summary: Validate required responses/evidence and complete a checklist
      security: *employeeWriteSecurity
      responses:
        <<: *standardErrors
        "200":
          description: Completed checklist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrderChecklistEnvelope"
...
  /organizations/{orgId}/photos/{photoId}:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/PhotoId"
    get:
      tags: [Photos]
      operationId: getPhoto
      summary: Get authorized photo metadata and a short-lived read URL
      responses:
        <<: *standardErrors
        "200":
          description: Photo metadata with optional short-lived read URL.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoEnvelope"
    patch:
      tags: [Photos]
      operationId: updatePhoto
      summary: Update photo caption
      security: *employeeWriteSecurity
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePhotoRequest"
      responses:
        <<: *standardErrors
        "200":
          description: Photo metadata updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoEnvelope"
    delete:
      tags: [Photos]
      operationId: deletePhoto
      summary: Soft-delete a photo when evidence rules permit
      security: *employeeWriteSecurity
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      responses:
        <<: *standardErrors
        "204":
          $ref: "#/components/responses/NoContent"
...
components:
  securitySchemes:
    employeeSession:
      type: apiKey
      in: cookie
      name: session
      description: |
        Logical HttpOnly, Secure, SameSite session cookie issued by the server-side
        Supabase Auth adapter. Provider access and refresh tokens are never exposed.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required for employee mutations in addition to the session cookie.
    lineSignature:
      type: apiKey
      in: header
      name: X-Line-Signature
      description: Base64 HMAC-SHA256 signature verified over the exact raw request bytes.
    workerBearer:
      type: http
      scheme: bearer
      bearerFormat: short-lived service JWT
      description: Audience-restricted workload identity; never accepted by employee routes.
    publicCapabilityBearer:
      type: http
      scheme: bearer
      bearerFormat: 256-bit opaque capability
      description: Raw public quote capability carried only in Authorization; never in the HTTP path or logs.
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      description: Organization tenant boundary checked against the active membership.
      schema:
        $ref: "#/components/schemas/Uuid"
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
    AssignmentId:
      name: assignmentId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
    PhotoId:
      name: photoId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag containing the last observed positive lock version.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
        examples: ['"3"']
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Scoped to actor, organization, method and route template.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]+$"
...
  headers:
    ETag:
      description: Strong aggregate lock-version ETag used by If-Match.
      required: true
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
...
    RequestId:
      description: Non-secret request correlation identifier.
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
  responses:
    NoContent:
      description: Operation completed with no response body.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
...
  schemas:
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
...
    AssignmentDuty:
      type: string
      enum: [lead, technician, helper, observer]
...
    ChecklistItemType:
      type: string
      enum: [boolean, text, number, single_choice, multi_choice, photo]
    ChecklistStatus:
      type: string
      enum: [pending, in_progress, completed, reopened]
    PhotoCategory:
      type: string
      enum: [intake, before, after, issue, receipt, signature, other]
    PhotoStatus:
      type: string
      enum: [pending, processing, ready, quarantined, failed, deleted]
...
    Assignment:
      type: object
      additionalProperties: false
      description: Assignment resource returned by create, update, cancel and respond.
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        organizationId:
          $ref: "#/components/schemas/Uuid"
        workOrderId:
          $ref: "#/components/schemas/Uuid"
        membershipId:
          $ref: "#/components/schemas/Uuid"
        duty:
          type: string
        status:
          type: string
        assignedAt:
          type: [string, "null"]
        acceptedAt:
          type: [string, "null"]
        declinedAt:
          type: [string, "null"]
        checkedInAt:
          type: [string, "null"]
        completedAt:
          type: [string, "null"]
        cancelledAt:
          type: [string, "null"]
        declineReason:
          type: [string, "null"]
        lockVersion:
          type: integer
      required: [id, organizationId, workOrderId, membershipId, duty, status, assignedAt, acceptedAt, declinedAt, checkedInAt, completedAt, cancelledAt, declineReason, lockVersion]
...
    UpdateAssignmentRequest:
      type: object
      additionalProperties: false
      properties:
        duty:
          $ref: "#/components/schemas/AssignmentDuty"
      required: [duty]
    CancelAssignmentRequest:
      type: object
      additionalProperties: false
      description: Reason is mandatory and carried in the body because DELETE has no natural payload.
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 2000
      required: [reason]
    RespondAssignmentRequest:
      type: object
      additionalProperties: false
      properties:
        decision:
          type: string
          enum: [accept, decline]
        reason:
          type: [string, "null"]
          maxLength: 1000
      required: [decision]
    AssignmentEnvelope:
      type: object
      additionalProperties: false
      properties:
        data:
          $ref: "#/components/schemas/Assignment"
      required: [data]
...
    ChecklistResponse:
      type: object
      additionalProperties: false
      properties:
        checked:
          type: [boolean, "null"]
        text:
          type: [string, "null"]
          maxLength: 5000
        number:
          type: [number, "null"]
        selectedOption:
          type: [string, "null"]
          maxLength: 200
        photoIds:
          type: array
          uniqueItems: true
          maxItems: 20
          items:
            $ref: "#/components/schemas/Uuid"
      required: [checked, text, number, selectedOption, photoIds]
    WorkOrderChecklistItem:
      type: object
      additionalProperties: false
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        label:
          type: string
          maxLength: 500
        itemType:
          $ref: "#/components/schemas/ChecklistItemType"
        required:
          type: boolean
        options:
          type: array
          maxItems: 100
          items:
            type: string
            maxLength: 200
        response:
          $ref: "#/components/schemas/ChecklistResponse"
        sortOrder:
          type: integer
          minimum: 0
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
      required: [id, label, itemType, required, options, response, sortOrder, updatedAt]
    WorkOrderChecklist:
      type: object
      additionalProperties: false
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        workOrderId:
          $ref: "#/components/schemas/Uuid"
        templateId:
          type: [string, "null"]
          format: uuid
        templateVersion:
          type: [integer, "null"]
          minimum: 1
        name:
          type: string
          maxLength: 200
        status:
          $ref: "#/components/schemas/ChecklistStatus"
        items:
          type: array
          maxItems: 200
          items:
            $ref: "#/components/schemas/WorkOrderChecklistItem"
        completedAt:
          anyOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        completedByMembershipId:
          type: [string, "null"]
          format: uuid
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
      required: [id, workOrderId, templateId, templateVersion, name, status, items, completedAt, completedByMembershipId, createdAt, updatedAt]
...
    WorkOrderChecklistEnvelope:
      type: object
      additionalProperties: false
      properties:
        data:
          $ref: "#/components/schemas/WorkOrderChecklist"
      required: [data]
...
    Photo:
      type: object
      additionalProperties: false
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        aggregateType:
          type: string
          enum: [service_request, work_order]
        aggregateId:
          $ref: "#/components/schemas/Uuid"
        category:
          $ref: "#/components/schemas/PhotoCategory"
        status:
          $ref: "#/components/schemas/PhotoStatus"
        filename:
          type: string
          minLength: 1
          maxLength: 200
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
        byteSize:
          type: integer
          minimum: 1
          maximum: 10485760
        width:
          type: [integer, "null"]
          minimum: 1
        height:
          type: [integer, "null"]
          minimum: 1
        caption:
          type: [string, "null"]
          maxLength: 1000
        capturedAt:
          anyOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        checklistItemId:
          type: [string, "null"]
          format: uuid
        downloadUrl:
          type: [string, "null"]
          format: uri
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
      required: [id, aggregateType, aggregateId, category, status, filename, contentType, byteSize, width, height, caption, capturedAt, checklistItemId, downloadUrl, createdAt, updatedAt]
...
    UpdatePhotoRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        category:
          $ref: "#/components/schemas/PhotoCategory"
        caption:
          type: [string, "null"]
          maxLength: 1000
        takenAt:
          anyOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
...
    PhotoEnvelope:
      type: object
      additionalProperties: false
      properties:
        data:
          $ref: "#/components/schemas/Photo"
      required: [data]
...